🩹 [Patch]: Update linter and git configurations and license year (#15)
## Description

This pull request includes several changes to configuration files for
linters and updates to the license file. The most important changes are
the addition of a new configuration file for `jscpd`, updates to the
PowerShell Script Analyzer configuration, and a minor update to the
license file.

Linter configuration updates:

*
[`.github/linters/.jscpd.json`](diffhunk://#diff-557094e283c00b23265c1c75872f41c6b1a524a00f0d99dd68ebd22cb63bfdd6R1-R10):
Added a new configuration file for `jscpd` with a threshold of 0,
consoleFull reporter, and exclusion of tests.
*
[`.github/linters/.powershell-psscriptanalyzer.psd1`](diffhunk://#diff-aae69c9d6774628ed181eacf53aee0f38eb6c2f53492cf3a5b7f7bdb6ef43b6aL1-L16):
Updated the configuration to enable specific rules for PowerShell script
analysis, including alignment, line length, semicolon usage, brace
placement, comment help, indentation, and whitespace consistency.
*
[`.github/workflows/Linter.yml`](diffhunk://#diff-482e65806ed9e4a7320f14964764086b91fed4a28d12e4efde1776472e147e79R30):
Disabled JSON Prettier validation in the Super Linter workflow.

License update:

*
[`LICENSE`](diffhunk://#diff-c693279643b8cd5d248172d9c22cb7cf4ed163a3c98c8a3f69c2717edd3eacb7L3-R3):
Updated the copyright year from 2024 to 2025.

## Type of change

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [ ] 📖 [Docs]
- [ ] 🪲 [Fix]
- [x] 🩹 [Patch]
- [ ] ⚠️ [Security fix]
- [ ] 🚀 [Feature]
- [ ] 🌟 [Breaking change]

## Checklist

<!-- Use the check-boxes [x] on the options that are relevant. -->

- [x] I have performed a self-review of my own code
- [x] I have commented my code, particularly in hard-to-understand areas